requirements.txt
# FastAPI core
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Migrations
alembic==1.13.1

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Auth / security
python-multipart==0.0.20
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy[sqlalchemy]==7.0.0
asyncpg==0.30.0

# Environment
python-dotenv==1.0.1

# HTTP requests
httpx==0.27.0
